Add repos interfaces, change the code to fit requirements

+++ src/main/java/com/library/RentRepository.java
@@ -1,9 +1,11 @@
 package com.library;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.UUID;
 
-public class RentRepository {
+public class RentRepository implements RentRepoInterface {
     private ArrayList<Rent> repository;
 
     public ArrayList<Rent> getRentRepository() {
@@ -14,12 +16,34 @@ public class RentRepository {
         this.repository = new ArrayList<>();
     }
 
-    public void add(Rent rent){
-        repository.add(rent);
+    public void add(Object rentObject) {
+        Rent rent = (Rent) rentObject;
+        for (Rent r : repository) {
+            if (r.getId() == rent.getId()) {
+                System.out.println("Istnieje już wypożyczenie o danym id");
+                return;
+            }
+            repository.add(rent);
+        }
     }
-
+// zakończenie alokacji polega na ustawieniu atrybutu czasu zakończenia alokacji
+    public void endRent(Rent rent){
+        for (Rent r : repository) {
+            if (r.getId() == rent.getId()) {
+                rent.setEndDate(LocalDate.now());
+            } else {
+                System.out.println("Brak podanego wypożyczenia");
+            }
+        }
+    }
+//usuwanie alokacji dotyczy tylko alokacji nie zakończonych
     public void remove(Rent rent){
-        repository.remove(rent);
+        if(rent.getEndDate() != null){
+            repository.remove(rent);
+        }
+        else {
+            System.out.println("Nie można usunąć zakończonego wypożyczenia");
+        }
     }
 
     @Override
@@ -40,10 +64,10 @@ public class RentRepository {
         return null;
     }
 
-    public ArrayList<Rent> getAllUserRents(User person){
+    public ArrayList<Rent> getAllUserRents(UUID id){
         ArrayList<Rent> userRents = new ArrayList<>();
         for (Rent rent : repository) {
-            if (rent.getUser() == person) {
+            if (rent.getUser().getId() == id) {
                 userRents.add(rent);
             }
         }
@@ -57,5 +81,18 @@ public class RentRepository {
             }
         }
         return "Brak podanego wypożyczenia w repozytorium";
+    }
+
+    public Rent getRentForElement(UUID id){
+        for(Rent rent : repository){
+            if(rent.getElement().getId() == id){
+                return rent;
+            }
+        }
+        return null;
+    }
+
+    public List<Rent> getRents(){
+        return repository;
     }
 }
